+++ src/amun/cpu/decoder/encoder_decoder.h
@@ -18,7 +18,7 @@ class CPUEncoderDecoderBase : public Scorer {
     	const God &god,
         const std::string& name,
         const YAML::Node& config,
-        size_t tab);
+        unsigned tab);
 
     virtual State* NewState() const;
