+++ src/amun/cpu/decoder/encoder_decoder_state.h
@@ -13,7 +13,7 @@ class EncoderDecoderState : public State {
     EncoderDecoderState();
     EncoderDecoderState(const EncoderDecoderState&) = delete;
 
-    virtual std::string Debug(size_t verbosity = 1) const;
+    virtual std::string Debug(unsigned verbosity = 1) const;
 
     CPU::mblas::Matrix& GetStates();
     const CPU::mblas::Matrix& GetStates() const;
